Decouple listening to a socket from awaitMessage
THere are cases where listening to a socket and waiting for new
connections in the socket are not tightly coupled. THerefore, we need to
decouple the listening part to a socket from waiting for a new
connection. An example of such a case is when we need to create a socket
(listen to a unix socket) before the creation of the reexec process.
For that purpose, this patch:
- Decouples the listening part from awaitMessage, creating a new
function
- Exposes awaitMessage directly through the unikontainers package
- Maintains the coupling of listening and accepting new connections to a
socket, by creating a universal function which takes as an argument the
expected message.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>