Introduce support to boot up Firecracker with intrd
Add initial support for booting unikernels over Firecracker with urunc.
For the time being, we do not have support for virtio-blk, but only for
initrd. THerefore, the guest can only read data from the initrd.
Regarding networking, we use virtio-net  over the tap interface  that we
create with urunc. Currently there is support for only one network
device per guest. Moreover, in the case of Firecracker the guest's
network interface has the same MAC address as the tap interface in the
network namespace of the container.

Overall, things we need to consider in the future:
- Support for virtio-blk
- VM memory
- Generate Firecracker's configuration, based on a specific config
- Use Firecracker's API to spawn, instead of creating a JSON file

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>