Add static and dynamic network modes, handle queue-proxy container for knative
This commit introduces a network manager struct to abstract the
specific network mode implementation. When unikernel container is a
knative user container, the static network is used. In every other
case, the dynamic network is used. In dynamic mode, multiple tap
interfaces in the same network namespace are now supported,
but only the first one provides functional networking.

Furthermore, to support knative we detect any queue-proxy containers and
inject a REDIRECT_IP ENV variable in the queue-proxy container process environment,
before calling runc to execute it. The REDIRECT_IP (172.16.1.1) is hardcoded for now.

Finally, the output file for the timestamp logging is defined
in internal/constants instead of a hardcoded string.

Signed-off-by: Georgios Ntoutsos <gntouts@nubificus.co.uk>