chore(docs): Update Linux tutorial to note that urunit is needed in K8s
When we deploy a Linux container with urunc in Kuberentes, we face the
issue where the gateway might be in a different subnet than the IP.
In such cases, Linux will fzil to set the default gateway and as such,
the VM will not have network access. Since urunc uses kernel boot
parameters to set up the network configuration of the VM, there is no
way to properly handle such issues.

As a reuslt, we cna only overcome this by an init process running inside
the VM. That is what urunit does, ehich sets the default route through
eth0. However, this makes necessary the use of urunit when we deploy
such containers in K8s. Therefore, we need to mention this in our tutorial.

With the opportunity remove the rumprun debug message for the JSON
construction.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>